requirements.txt
# 核心依赖
langchain
langchain-openai
openai
python-dotenv
langchain-community
pydantic
